requirements.txt
# requirements.txt
llama-cpp-python
llama-stack
sentence-transformers
tqdm
faiss-cpu
python-docx
PyPDF2
pytesseract
transformers
torch
fuzzywuzzy
python-Levenshtein
easyocr
hf_xet
pdf2image
PyYAML
pillow
streamlit
nltk
scikit-learn
spacy
safetensors